修正

+++ install.sh
@@ -4,12 +4,28 @@ set -eu
 cd ~/dotfiles
 
 # mysql
-ln -s ~/dotfiles/my.cnf ~/.my.cnf
+if [ -e ~/.my.cnf ]; then
+    rm ~/.my.cnf
+    ln -s ~/dotfiles/my.cnf ~/.my.cnf
+fi
 
 # tmux
-ln -s ~/dotfiles/tmux.cnf ~/.tmux.conf
+if [ -e ~/.my.tmux.conf ]; then
+    rm ~/.my.tmux.conf
+    ln -s ~/dotfiles/tmux.cnf ~/.tmux.conf
+fi
 
 # vim
 git clone https://github.com/enoatu/vimfiles.git
-./vimfiles/install.sh
+cd vimfiles
+chmod +x ./install.sh
+./install.sh
 
+if [ -e ./vimfiles ]; then
+    rm -rf ./vimfiles
+fi
+
+sed -i '/let s:dein_dir/s/~\//~\/dotfiles\//' vimrc
+ln -s ~/dotfiles/vimfiles/vimrc ~/.vimrc
+
+printf "\e[30;42;1dotfiles setup completed\e[m\n"
